+++ test/process.js
@@ -17,12 +17,10 @@
 // You should have received a copy of the GNU General Public License
 // along with this nabs.  If not, see <http://www.gnu.org/licenses/>.
 
-'use strict';
-
 const chai = require('chai');
 const td = require('testdouble');
 
-const nabs = require('../dist/nabs');
+const nabs = require('../dist/nabs').default;
 
 chai.should();
