+++ webpack/webpack.common.js
@@ -1,16 +1,17 @@
-'use strict';
-
 const webpack = require('webpack');
-const { resolve } = require('path');
+const { resolve, join } = require('path');
+const { readFileSync } = require('fs');
 const NodeExternals = require('webpack-node-externals');
 
 const src = resolve('src');
 const dist = resolve('dist');
+const banner = readFileSync(resolve('webpack/file-header.txt')).toString().concat('\n');
 
 module.exports = {
   target: 'node',
   entry: {
-    nabs: src,
+    index: src,
+    nabs: join(src, 'nabs.js'),
   },
   output: {
     filename: '[name].js',
@@ -28,7 +29,14 @@ module.exports = {
     ],
   },
   plugins: [
-    new webpack.BannerPlugin({ banner: '#!/usr/bin/env node', raw: true }),
+    new webpack.BannerPlugin({
+      banner,
+      raw: true,
+    }),
+    new webpack.BannerPlugin({
+      banner: '#!/usr/bin/env node\n',
+      raw: true,
+    }),
   ],
   externals: [
     NodeExternals(),
